agent/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
httpx>=0.25.0
pydantic>=2.0
python-dotenv>=1.0.0
